Refactor Analyzer and PrivilegeChecker visitor to support Dependency Injection
